Simplify KeycloakRemoteUserProvider and fix closed keycloakClient error
All KeyCloak results are lists, so there is not much point in returning streams at the moment. Above all, this brought with it the problem that with the groups, for example, the keycloakClient could already be closed by the auto-close even though not all groups were loaded yet...of course, this could have been solved differently, but since keycloakClient does not support streams with these methods anyway, it is cleaner this way.