Introduce a soft limit regarding license seat count used during unlock
We now have a hard limit used when adding new members to a vault or
granting access permission. Here the effective members are counted.

In addition to that, this commits adds a soft limit which is used during
unlock. Only users that have at least one token does count against this
limit. Now we can still unlock the vaults when through a group the seat
count is higher than available seats in the license. Users can not be
granted permission or added to a vault but existing users can still
unlock those.