Fix license refresh (#341)
`skipExecutionIf` runs in its own thread, so there is no access to the actual beans in this case. This leads to a `jakarta.enterprise.context.ContextNotActiveException` (runtime exception), while calling (again) init in the `LicenseHolder` which crashes the thread and leads to `licenseHolder.license` is always null in this `Scheduled.SkipPredicate` environment.

This reverts the changes made in #266 concerning the handling of licence refreshes.